drm/panic: Add a drm panic handler
This module displays a user friendly message on screen when a kernel panic
occurs. As the kernel already panicked, don't try to change screen
resolution, just grab the framebuffer, and overwrite it if possible.
It will display only one frame, so just clear it, and draw white pixels for
the characters. Performance optimizations are low priority as the machine
is already in an unusable state.

Signed-off-by: Jocelyn Falempe <jfalempe@redhat.com>