drm/panic: Add a qr_code panic screen
This patch adds a new panic screen, with a qr code and the kmsg
data embedded. Distribution can customize the url, so they can put a
web frontend to directly open a bug report with the kmsg data.
If DRM_PANIC_SCREEN_QR_CODE_URL is not set, the kmsg data will be
written so you can copy/paste it from your smartphone.

The qr encoder is self-contained and written in rust.
kmsg is first compressed with zlib, and then encoded as a numerical
segment in the qr-code, and appended to the url as url parameter.

The biggest QR-code is V40 (177x177) but some phones have problems
reading QR-code bigger than V23 (109x109), so that limits the kmsg
data that can be encoded inside.
This can be customized with DRM_PANIC_SCREEN_QR_KMSG_SIZE, at build
time or with drm.panic_qr_kmsg_size=2048 at boot.

Signed-off-by: Jocelyn Falempe <jfalempe@redhat.com>